Create Unit test projects and dispose at test closure

+++ Clarizen.Tests/Steps/Api_UserSteps.cs
@@ -19,7 +19,7 @@ namespace Clarizen.Tests.Steps
 
             var results = Context.Api.ExecuteQuery(query);
             Assert.True((results.Error == null), results.Error);
-            Assert.Equal(expectedAdminUserCount, results.Data.entities.Length);
+            Assert.Equal((int)expectedAdminUserCount, (int)results.Data.entities.Length);
         }
     }
 }
